백엔드 구조 개선
- 모델 예측 관련 endpoint를 main.py에서 router/model.py로 분리하여 구조화
- 저장된 전처리 이미지를 .gitignore 처리
